refactor it
