+++ components/card/demo/_semantic.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import SemanticPreview from '../../../.dumi/theme/builtins/SemanticPreview';
+import SemanticPreview from '../../../.dumi/theme/common/SemanticPreview';
 import useLocale from '../../../.dumi/hooks/useLocale';
 
 import { EditOutlined, EllipsisOutlined, SettingOutlined } from '@ant-design/icons';
